refactor(dev-tool): split compiler tests into smaller files
Extract self-contained test sections from compiler.spec.ts:
- helperCompilation.spec.ts: tests for helper file compilation (285 lines)
- generateSampleEnv.spec.ts: tests for env file generation (62 lines)

Reduces main compiler.spec.ts from ~2190 to 1852 lines.
